Refactor: NotionAPI 데이터 받아오는 방법 변경 [#23]
